Move bot files into bot directory
