+++ bot/assets/checks.py
