+++ bot/assets/excuses.txt
