+++ bot/assets/words.py
