+++ bot/cogs/codesandbox.py
