+++ bot/cogs/converters.py
