+++ bot/cogs/embeds.py
