+++ bot/cogs/emote_utils/emotes.py
