+++ bot/cogs/emote_utils/exceptions.py
