+++ bot/cogs/emotes.py
