+++ bot/cogs/endpoints/endpoints.py
