+++ bot/cogs/search.py
