+++ bot/cogs/search_utils/regex.py
