+++ bot/cogs/search_utils/searchexceptions.py
