+++ bot/cogs/translate.py
