+++ bot/cogs/utils/errors.py
