+++ bot/utils/binarycalc.py
