+++ bot/utils/hexcalc.py
