+++ bot/utils/mathscrape.py
