+++ start.py
@@ -1,0 +1,4 @@
+from bot import __main__ as main
+
+if __name__ == "__main__":
+    main()
